third_party/WebKit/LayoutTests/fast/events/middleClickAutoscroll-nested-divs-forbidden.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gesture-util.js"></script>
<script src="../../resources/compositor-controls.js"></script>
<script src="./resources/middleClickAutoscroll.js"></script>
<style type="text/css">
div {
  line-height: 50px;
}
#outerdiv {
  overflow:auto;
  border: 2px solid #000FFF;
}
#innerdiv {
  overflow:auto;
  border: 2px solid #000000;
  width: 121%;
  padding: 5px;
}
</style>
<script>
function start() {
  outerdiv = $('outerdiv');

  testSetUp({
    'clickOrDrag': 'drag',
    'endX': 225,
    'endY': 75,
    'finishTest': function() {
      assert_not_equals(outerdiv.scrollLeft, 0);
      assert_equals(outerdiv.scrollTop, 0);
    },
    'scrollable': outerdiv,
    'startX': 150,
    'startY': 150,
  });
}
</script>
<body onload="start()">
<div id="outerdiv">
  <p>Top of outer div.</p>
  <div id="innerdiv">
    <p>Inner div.</p>
  </div>
  <p>Bottom of outer div.</p>
</div>
<p>Test for <a href="http://crbug.com/232965">bug 232965</a> This tests that vertical pan scrolling does not propagate from the inner div to the outer div when the outer div has no vertical overflow.</p>
<div id="console"></div>
</body>
